+++ CHANGELOG.md
@@ -10,7 +10,11 @@
 - improve UI: faster way to manage data-table records (targets/groups/tests) by keeping the last row always in editing more
 - improve UI: suite options moved under a tab and do not take space in main workspace
 - improve UI: Suite panes menu (Target/Tests/...) set for sticky position (always accessible regardless of scrolling)
-- improve UI: added Insert action for data table records, which injects an new empty record next to the selected one set editing mode
+- improve UI: add Insert action for data table records, which injects an new empty record next to the selected one set editing mode
 - improve UI: integration with GIT (local/remote)
 - improve UI: added copy/paste. Now one can copy into clipboard target/group/test/command in one suite/GIT version and paste on another
 - improve UI: Project Explorer gets support for multiple project
+- improve UI: add dynamic environment variables
+- improve UI: add templates
+- improve UI: add command assignVarRemotely
+- improve UI: add command assertVar